Introduces era parmetric Fields, and updaters
The parmetric fields can be used with the updaters to apply a field to
an appropriate objects from any era.  Also introduced is the Policy type
that controls how the value in the field is used to update the object.

Using function calls like: newXXX policy eraWitness [field1,..,fieldn]
one can create new objects where the missing fields are given default
values.